changed assets folder location

+++ src/ecs/systems/AABBSystem.hpp
@@ -15,8 +15,8 @@ struct AABBSystem : public BaseSystem
 			Transform* t1 = ecs.GetComponent<Transform>(e1);
 			
 			// Calculate new boundaries for box1
-			box1->min = t1->position.subtract(box1->dimensions.multiply(t1->scale).divide(2));
-			box1->max = t1->position.add(box1->dimensions.multiply(t1->scale).divide(2));
+			box1->min = t1->position.subtract(box1->dimensions.multiply((float)t1->scale).divide(2));
+			box1->max = t1->position.add(box1->dimensions.multiply((float)t1->scale).divide(2));
 
 			for (auto it2 = std::next(it1); it2 != entities.end(); ++it2)
 			{
@@ -26,8 +26,8 @@ struct AABBSystem : public BaseSystem
 				Transform* t2 = ecs.GetComponent<Transform>(e2);
 
 				// Calculate new boundaries for box2
-				box2->min = t2->position.subtract(box2->dimensions.multiply(t2->scale).divide(2));
-				box2->max = t2->position.add(box2->dimensions.multiply(t2->scale).divide(2));
+				box2->min = t2->position.subtract(box2->dimensions.multiply((float)t2->scale).divide(2));
+				box2->max = t2->position.add(box2->dimensions.multiply((float)t2->scale).divide(2));
 
 				// Check for collisions between 2 bounding boxes
 				box1->check(*box2);
